Add some entities to the task solution
