+++ src/build-logic/src/main/kotlin/plugins/kmp-library-plugin.gradle.kts
@@ -1,12 +1,11 @@
 @file:Suppress("UnstableApiUsage")
 
-import com.android.build.gradle.internal.tasks.factory.registerTask
 import config.ConfigurationKeys
 import ext.configureCompileOptions
 import ext.configureDefaultConfig
 import ext.configurePlatformTargets
 import ext.configureTestOptions
-import tasks.CheckArchitectureViolationTask
+import tasks.checkarcviolation.CheckArchitectureViolationTask
 
 @Suppress("DSL_SCOPE_VIOLATION") // TODO: Remove once KTIJ-19369 is fixed
 plugins {
